+++ src/0144.Binary-Tree-Preorder-Traversal/Solution.go
@@ -31,33 +31,3 @@ func preorderTraversal(root *TreeNode) []int {
 
 	return nodes
 }
-
-const maxNum = 100
-
-type seqStack struct {
-	data [maxNum]*TreeNode
-	tag  [maxNum]int // 后续遍历准备
-	top  int         // 数组下标
-}
-
-func preorderTraversal2(root *TreeNode) []int {
-	var s seqStack
-	s.top = -1
-	result := []int{}
-
-	if root == nil {
-		return result
-	}
-	for root != nil || s.top != -1 {
-		for root != nil {
-			result = append(result, root.Val)
-			s.top++
-			s.data[s.top] = root
-			root = root.Left
-		}
-		s.top--
-		root = s.data[s.top+1]
-		root = root.Right
-	}
-	return result
-}
